Initial Commit of new CSS based game
The game prompts the user to style a specific element accordingly
giving the user 30 seconds for each task.  Once the user correctly
styles the element, the user is tasked with a new style to apply.
Points are awarded after every correct style application.  The time is
reset to 30 seconds after every correct styling.  The user loses when
time runs out!